Add support for key multiplexing
Add multi-tenancy as an operational consideration. The idea is that if
multiple servers share the same IP or port space, then the client needs
a way to indicate which server it expects to respond to a request.

To resolve this, define a new request tag, SRV, that indicates the
long-lived public key the client expects to verify the response with.